Action関数の共通化
